[OCT-598] feat: add gestion of access for part video(delete and modify video) (#618)
* change any to pass the pre-commit

* feat: change the message to french

* fix: now we have 2 different language for toast when users are trying to delete or modify the videos that does not belong to them

* feat: now the button modify does not appear in detail page when user are not the owner of video

---------

+++ backend/src/videos/domain/usecases/deleteVideo.usecase.ts
@@ -1,15 +1,35 @@
-import { Inject, Injectable } from '@nestjs/common';
+import {
+  Inject,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { IVideoGateway } from '../gateways/videos.gateway';
 import { VideoObject } from '../video';
+import { IUserGateway } from 'src/users/domain/gateways/users.gateway';
 
 @Injectable()
 export class DeleteVideoUsecase {
   constructor(
     @Inject('VideoGateway')
     private videoGateway: IVideoGateway,
+    @Inject('UserGateway')
+    private userGateway: IUserGateway,
   ) {}
 
-  async execute(id: string): Promise<VideoObject> {
+  async execute(id: string, email: string): Promise<VideoObject> {
+    const user = await this.userGateway.getUserByEmail(email);
+    if (!user) throw new NotFoundException(`User (email: ${email}) not found`);
+
+    const videos = await this.videoGateway.getVideos({ id });
+    if (!videos || videos.length === 0)
+      throw new NotFoundException(`Video (id: ${id}) not found`);
+    const video = videos[0];
+
+    if (!video.creator || video.creator.email !== user.email) {
+      throw new UnauthorizedException('videonotallowdeleted');
+    }
+
     return await this.videoGateway.deleteVideo(id);
   }
 }
